Set release candidate version suffix and bump SOVERSION
Change the version suffix from -devel-{TODAY} to -rc

Bump the SO version for libclamav and libfreshclam.
Increasing the current version to 12.0.0 and 3.0.0 respectively.
The reason is that we reintroduced using the version scripts:
- libclamav.map
- libfreshclam.map
- libclamunrar.map
- libclamunrar_iface.map

Note that libclamunrar and libclamunrar_iface use the SO version from
libclamav.

Note that libclammspack does not have a .map file and so is not getting
symbol versioning at this time nor are we bumping the SO version for
that library.