Windows: Enable MSI installer to upgrade between feature versions
The MSI installer can find previous ClamAV installs to non-standard
locations such at D:\ClamAV instead of C:\Program Files\ClamAV but only
if upgrading to a new patch version of the same feature version.

E.g. can find an upgrade 1.0.0 to 1.0.1, but cannot automatically find
and upgrade 1.0.1 to 1.1.0.

This change will make it so all minor feature versions will store the
install path to the same key in the registry.

A consequence of this change is that you can no longer install multiple
feature versions of ClamAV using the MSI installer.
If you want multiple ClamAV minor versions installed on Windows you will
need to install using the ZIP package.

This change will not allow the MSI installer to automatically find and
upgrade across different major versions.
E.g. cannot find and upgrade 0.105.2 to 1.1.0 because the former is
major version 0, and the latter is major version 1.
This is intentional because it is more likely that there will be
breaking changes to config files and other user interfaces when we go to
ClamAV 2.0.0.