Freshclam, Sigtool: Fix bug creating new files in CLD with CDIFF
The CLOSE command is failing to create a file when appending changes if
the file does not already exist. This prevents adding new files to a
database with a CDIFF and caused failures applying the test-3.cdiff file
in the freshclam feature tests.

Also improved the error message to show which command, specifically, is
failing (not just the line number).