Resolve Coverity assignment of overlapping memory warnings
Coverity is unhappy with the use of the EC32, cli_readint32,
and cli_writeint32 macros (and the 64bit equivalents to potentially
change the endianess of variables in place.
It claims:

overlapping_assignment: Assigning ... to ..., which have overlapping
memory locations and different types.

Using a temporary variable in between reading and writing should
resolve these "high impact" complaints.

Resolves: Coverity-225232. 225225, 225215, 225212, 225180, 225170,
225165, 225161, 225159.