Jenkins: Clear cache instead of updating lockfile.
The intention behind adding `cargo update` in our Jenkinsfile earlier
was to make it so the local cache of the cargo registry index (i.e.
$HOME/.cargo/registry/index) knows about the latest version of the Rust
libraries required by our Cargo.lock file.

As it turns out, `cargo update` modifies the Cargo.lock file, which
was not intended.

This change makes removes use of `cargo update`.

It appears the real issue is that we needed a newer version of Rust to
build the 'image' crate.