Jenkins: Clear cache instead of updating lockfile.
The intention behind adding `cargo update` in our Jenkinsfile earlier
was to make it so the local cache of the cargo registry index (i.e.
$HOME/.cargo/registry/index) knows about the latest version of the Rust
libraries required by our Cargo.lock file.

As it turns out, `cargo update` may not actually update that index
(don't know why, exactly) but it DOES modify the Cargo.lock file, which
was not intended.

This change makes removes use of `cargo update`.
On our own time, if we need to, we can use the `cargo-cache` tool
or else manually clear the $HOME/.cargo/registry cache so the next build
will get the latest index when it builds. Thus it will know of, and
download, the correct sources as required by the Cargo.lock file.