OLE2: Fix bounds check on OLE2 encryption info check
The checks for the encryption info cspName and encryption verifier don't
have the size of the overall file available for the check and may
overflow.

This commit passes in the size of the file to the
initialize_encryption_key() function and does all size checks within
that function instead of doing the overall size check before that
function.

Resolves: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=60563