ClamD: Add new zOPTSCAN command w. scan options like ALLMATCH
Added a new ClamD command which can pass options along with the actual
scan command.

Current options are just `ALLMATCH`.

Current scan commands are:
- `CONTSCAN`
- `FILDES`
- `INSTREAM`
- `MULTISCAN`

For example, you can send:

  zOPTSCAN CONTSCAN /some/file/path\0

or:

  zOPTSCAN ALLMATCH INSTREAM \0
  <4-byte length><length-bytes data chunk>
  <4-byte length><length-bytes data chunk>
  <4-byte zero>

Note that a space is presently required between the final option
before the path or before the \0 for INSTREAM and FDPASS where the
subsequent bytes appear are sent in the next message.