Specify RuntimeDirectory for services
Specify RuntimeDirectory for clamav services `/run/clamav` to make
PIDFiles writeable.

The RuntimeDirectory ownership is changed by SystemD to match the `User`
and `Group` specified in the service unit files. ClamOnAccess runs as
root and therefore would clobber the ownership of these directories set
by other services in the family. For this reason, until a better
approach is available, RuntimeDirectory and LogsDirectory are not
managed by SystemD for ClomOnAccess service.