libclamav: resolve macOS linker issue with new LHA/LZH feature
The delharc crate used to add LZH archive support appears to add
a dependency on macOS CoreFoundation library.

The error is:

[ 78%] Linking C shared library libclamav.dylib
Undefined symbols for architecture x86_64:
  "_CFRelease", referenced from:
      iana_time_zone::platform::get_timezone_inner::hc7da204717a39974 in libclamav_rust.a(iana_time_zone-bc4762a47da73d72.iana_time_zone.1863eb20d202562a-cgu.0.rcgu.o)
...
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [libclamav/libclamav.12.0.2.dylib] Error 1

We already link with CoreFoundation for libfreshclam and clamsubmit, so
this commit extends that to libclamav as well.