Windows: fix build issue with newer version of Rust
Newer versions of Rust appear to include '/defaultlib:msvcrt' in
the list of native static libs when you run '--print=native-static-libs'.

MSVC doesn't know what to do with this when passed as a linker dependency.

Since msvcrt.lib is ALSO listed, the solution seems to be to strip this
item from the list.