Added cryptographic method flexibility to external signatures.  This includes a demo of post-quantum encryption using the open-quantum-safe project providers for OpenSSL v3 and later.
You must have liboqs (https://github.com/open-quantum-safe/liboqs) installed, and the oqsprovider for OpenSSL v3 and later (https://github.com/open-quantum-safe/oqs-provider).  Please note that you will also need to have altered the openssl.cnf file to include the oqsprovider in your configuration.  Instructions are on the oqsprovider page.

I have tested dilithium2 quantum resistant signatures and have included the source in a way that it will not cause faults if the provider is not installed.

It should be noted that at the present time NIST has NOT approved dilithium2 for use in FIPS systems.  I have put hard blockers on MD5 and SHA1 when running in FIPS mode to prevent mishaps.

The signing bash script will now hash and sign files in a flexible manner.  Algorithms support for SHA256 and SHA3-256 have been added to the existing hashing functions of the library.

I have validated functionality and ran valgrind against this commit.  I can find no bugs, but would always love to have a second or third set of eyes to help spot what I missed.

Shell script name changed to be more generic in the spirit of supporting arbitrary hashing and signing algorithms.  Script is now called:
clamav/sigext/cvd_ext_sigh.sh

If needed, I can add Elliptical Curve keys and signatures, as well, but it should be fairly elementary for anyone to do so now.