Fix bounds check in OLE2 decryption
The bounds check for the loop iterating an OLE2 block during decryption
may have an integer unerflow if the `leftover + bytesToWrite` is less
than 16. That results in a significant buffer over read and a segfault.

The fix is simply to do addition on the left side of the check instead
of subtraction on the right.

Fixes https://issues.oss-fuzz.com/issues/372544101